Clear all messages from previous views in a single step.
Otherwise, there is a high likelihood of the generated behaviors ending in a suffix of Timeout and DiscardMessage action.  This situation arises when there is no primary replica, and Timeout and DiscardMessage are the only enabled actions. For the BecomePrimary action to become enabled for some replicas, multiple DiscardMessage actions must occur. However, the simulation may frequently schedule a Timeout action in between, leading to additional DiscardMessage steps, prolonging the process unnecessarily.

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>